+++ main.c
@@ -10,6 +10,5 @@ int main(int argc, char* argv[])
   safire("Top-Down Shooter game", 960, 540, false, scenes, scenes_count);
   // safire("Top-Down Shooter game", 1920, 1080, false, scenes, scenes_count);
 
-  
   return 0;
 }