+++ topdown_shooter/include/topdown_shooter/bullet.h
@@ -1,6 +1,10 @@
 #ifndef __TDS_BULLET_H__
 #define __TDS_BULLET_H__
 
+#ifdef __cplusplus 
+extern "C" {
+#endif 
+
 #include <safire/safire.h>
 
 #define TDS_BULLET                      "Bullet"
@@ -26,5 +30,9 @@ SFR_Component*                          tds_bullet(vec2 move_direction, const ch
 
 
 
+
+#ifdef __cplusplus
+}
+#endif 
 
 #endif // __TDS_BULLET_H__