+++ topdown_shooter/include/topdown_shooter/camera_controller.h
@@ -1,6 +1,10 @@
 #ifndef __TDS_CAMERA_CONTROLLER_H__
 #define __TDS_CAMERA_CONTROLLER_H__
 
+#ifdef __cplusplus 
+extern "C" {
+#endif
+
 #include <safire/safire.h>
 
 #define TDS_CAMERA_CONTROLLER           "camera controller"
@@ -19,5 +23,8 @@ SAFIRE_API SFR_Component*               tds_camera_controller();
 
 
 
+#ifdef __cplusplus
+}
+#endif 
 
 #endif // __TDS_CAMERA_CONTROLLER_H__