+++ topdown_shooter/include/topdown_shooter/tile.h
@@ -1,6 +1,10 @@
 #ifndef __TDS_TILE_H__
 #define __TDS_TILE_H__
 
+#ifdef __cplusplus 
+extern "C" {
+#endif
+
 #include <safire/safire.h>
 
 #define TDS_TILE                        "tile"
@@ -14,5 +18,9 @@ SAFIRE_API SFR_Component*               tds_tile(vec2 location);
 
 
 
+
+#ifdef __cplusplus
+}
+#endif 
 
 #endif // __TDS_TILE_H__