+++ topdown_shooter/scenes/scene_arena.h
@@ -6,8 +6,8 @@
 
 
 // world size
-#define TDS_ARENA_TILES_X               40
-#define TDS_ARENA_TILES_Y               40
+#define TDS_ARENA_TILES_X               100
+#define TDS_ARENA_TILES_Y               100
 // non walkable space in world
 #define TDS_ARENA_TILES_WALLS_X         5  // non walkable space but still visable at the edge of the world
 #define TDS_ARENA_TILES_WALLS_Y         5